% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_smc.R
\name{smcSampler}
\alias{smcSampler}
\title{SMC Sampler for Bayesian Simulated Annealing}
\usage{
smcSampler(
  logLikelihoodDensity,
  logPriorDensity,
  priorSampler,
  numSamples = 200,
  numDistributions = 500,
  dimParameter,
  datum,
  ...
)
}
\arguments{
\item{logLikelihoodDensity}{A function that returns the log denisty of the
likelihood.}

\item{logPriorDensity}{A function that retunrs the log denisty of the prior.}

\item{priorSampler}{A functoin that returns samples from the prior.}

\item{numSamples}{A number specifiying the number of particles to be used in
the SMC sampler.}

\item{numDistributions}{A number specifiying the number of intermediate
distributions to be used in the simulated annealing (when adpatice CESS is
not used).}

\item{dimParameter}{The dimension of the paramaters space.}

\item{datum}{A vector containing the data point.}

\item{...}{Other varaibles to be passed onto likelihood, prior etc.}
}
\description{
\code{smcSamplerGeneral} Returns a List containing various values of the SMC
sampler e.g.  particle and (log) weight collection, normalising constant of
posterior denisty and other adaptive techniques metadata.
}
\details{
This is a general function that will provide weighted samples that can be used
to approximate different properties of the Bayesian model of interest
}
\examples{

\dontrun{
smcOutput = smcSampler(logLikelihoodDensity = petLogLikelihoodDensity,
logPriorDensity = petLogBioPriorDensity,
priorSampler = petBioPriorSampler,
dimParameter = 5,
datum = data1,
timeFrames = timeFrames)
}

}
