% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smcSamplerCpp}
\alias{smcSamplerCpp}
\title{Cpp SMC Sampler Implementation (PET ONLY)}
\usage{
smcSamplerCpp(
  numParticles,
  numOfDistrubtions,
  dataSet,
  timeFrames,
  dimParameters
)
}
\arguments{
\item{numParticles}{A number specifiying the number of particles to be used in
the SMC sampler.}

\item{timeFrames}{A vector containing the data point.}

\item{numOfDistributions}{A number specifiying the number of intermediate
distributions to be used in the simulated annealing (when adpatice CESS is
not used).}

\item{dataset}{A vector containing the data point.}

\item{dimParameter}{The dimension of the paramaters space.}
}
\description{
IMPORTANT: Recommended that you use \code{smcSamplerRcpp} instead !!
IMPORTANT: This function is used for development and unit testing only!!
\code{smcSamplerCpp} Returns a List containing various values of the SMC
sampler e.g.  particle and (log) weight collection, normalising constant of
posterior denisty and other adaptive techniques metadata.
}
\details{
This is NOT a general function  like the other samplers, instead this works only
for the PET model. However, like \code{smcSampler}, it will provide weighted
samples that can be used to approximate different properties of the Bayesian
model of interest
}
