% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_smc.R
\name{smcSamplerRcpp}
\alias{smcSamplerRcpp}
\title{SMC Sampler for (PET Model) - Rcpp Implementation}
\usage{
smcSamplerRcpp(
  numSamples = 200,
  numDistributions = 500,
  datum,
  timeFrames,
  dimParameter
)
}
\arguments{
\item{numSamples}{A number specifiying the number of particles to be used in
the SMC sampler.}

\item{numDistributions}{A number specifiying the number of intermediate
distributions to be used in the simulated annealing (when adpatice CESS is
not used).}

\item{datum}{A vector containing the data point.}

\item{dimParameter}{The dimension of the paramaters space.}
}
\description{
IMPORTANT: This function only works for the PET model!
Returns a List containing various values of the SMC
sampler e.g.  particle and (log) weight collection, normalising constant of
posterior denisty and other adaptive techniques metadata.
}
\details{
This is NOT a general function. It will return provide weighted samples that
can be used to approximate different properties of the PET Bayesian model.
}
\examples{

\dontrun{
smcOutput = smcSamplerRcpp( dimParameter = 5, datum = data1,
timeFrames = timeFrames)
}

}
